# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/contests/{contestId}/scoreboard:
    get:
      tags:
      - scoreboard-controller
      operationId: getPublicScoreboard
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ScoreboardResponse"
...
  /api/contests/{contestId}/scoreboard/admin:
    get:
      tags:
      - scoreboard-controller
      operationId: getAdminScoreboard
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ScoreboardResponse"
...
  /api/admin/contests/{contestId}/exports/scoreboard.csv:
    get:
      tags:
      - data-export-controller
      operationId: scoreboard
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    FirstBloodEntry:
      type: object
      properties:
        problemId:
          type: integer
          format: int64
        problemAlias:
          type: string
        teamId:
          type: integer
          format: int64
        teamName:
          type: string
        submittedAt:
          type: string
          format: date-time
...
    ScoreboardCell:
      type: object
      properties:
        problemId:
          type: integer
          format: int64
        alias:
          type: string
        solved:
          type: boolean
        attempts:
          type: integer
          format: int32
        wrongAttempts:
          type: integer
          format: int32
        firstAcceptedAt:
          type: string
          format: date-time
        penaltyMinutes:
          type: integer
          format: int64
        firstBlood:
          type: boolean
    ScoreboardProblem:
      type: object
      properties:
        problemId:
          type: integer
          format: int64
        alias:
          type: string
        name:
          type: string
    ScoreboardResponse:
      type: object
      properties:
        contestId:
          type: integer
          format: int64
        contestName:
          type: string
        variant:
          type: string
        groupName:
          type: string
        frozen:
          type: boolean
        freezeAt:
          type: string
          format: date-time
        generatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
        problems:
          type: array
          items:
            $ref: "#/components/schemas/ScoreboardProblem"
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ScoreboardRow"
        firstBlood:
          type: array
          items:
            $ref: "#/components/schemas/FirstBloodEntry"
    ScoreboardRow:
      type: object
      properties:
        rank:
          type: integer
          format: int32
        teamId:
          type: integer
          format: int64
        teamName:
          type: string
        school:
          type: string
        star:
          type: boolean
        participationType:
          type: string
        groupName:
          type: string
        solved:
          type: integer
          format: int32
        penaltyMinutes:
          type: integer
          format: int64
        latestAcceptedAt:
          type: string
          format: date-time
        cells:
          type: array
          items:
            $ref: "#/components/schemas/ScoreboardCell"
...